refactor: Remove CredentialStore
